Version 1.1 of the LSD mount utilities: Support mounting/unounting encrypted devices.
This was the first public release from around 2007-01-23. I think I remember that 1.0 was
an internal release that had been in use some time. It can mount and unmount encrypted
file systems, but is lacking LUKS, fsck and mkfs support.

This commit encompasses all files from the archive lsd-mount-utilities-1.1.tar.gz which
has the following checksums:

- sha512: 14c84d60f2bc0d4727d5e56948128ffc0196201d1982e6267b92715b3db40ed0d4de8348571f7d79db80237031c7a20d29b2d6bec2f5a108d6ec13d39f62a356
- md5sum: 3229c88b0a2687bbb359feea7da0c013